+++ NEWS.md
@@ -1,3 +1,9 @@
+# tarflow.iquizoo 0.1.2
+
+## New Features
+
+* Supported games abilities.
+
 # tarflow.iquizoo 0.1.1
 
 * Adapt to new database design in IQUIZOO.